Fog/requirements.txt
paho-mqtt
flask
flask_socketio
tensorflow
scikit-learn
numpy
h5py